Added configs entry, tweaked Trove JS
JavaScript for Trove entry links now takes an argument of the repo/filename, rather than a boolean for if it is in the trove and relying on the title in the entry's HTML file. For example, for my repo 'configs', 'configs' is the argument given to createGithubLink() and will be appended to "https://www.github.com/bradendubois/".